refactor tailRuns to allow programmatic execution
